+++ src/checkUsageCalc.ts
@@ -1,10 +1,7 @@
-import { CheckType } from './types';
-
 interface ActiveSeriesParams {
   probeCount: number;
-  checkType: CheckType;
   frequencySeconds: number;
-  useFullMetrics: boolean;
+  seriesPerCheck: number;
 }
 
 interface UsageValues {
@@ -13,43 +10,11 @@ interface UsageValues {
   logsGbPerMonth: number;
 }
 
-enum CheckSeries {
-  PING = 76,
-  HTTP = 128,
-  DNS = 79,
-  TCP = 59,
-  Traceroute = 20,
-}
-
-enum CheckSeriesBasic {
-  PING = 25,
-  HTTP = 38,
-  DNS = 28,
-  TCP = 23,
-  Traceroute = 20,
-}
-
-const getSeriesPerCheck = (checkType: CheckType, useFullMetrics: boolean) => {
-  const Series = useFullMetrics ? CheckSeries : CheckSeriesBasic;
-  switch (checkType) {
-    case CheckType.PING:
-      return Series.PING;
-    case CheckType.TCP:
-      return Series.TCP;
-    case CheckType.DNS:
-      return Series.DNS;
-    case CheckType.HTTP:
-      return Series.HTTP;
-    case CheckType.Traceroute:
-      return Series.Traceroute;
-  }
-};
-
 const getChecksPerMonth = (frequencySeconds: number) => {
-  const checksPerMinute = Math.round(60 / frequencySeconds);
+  const checksPerMinute = 60 / frequencySeconds;
   const checksPerHour = checksPerMinute * 60;
   const checksPerMonth = checksPerHour * 730;
-  return checksPerMonth;
+  return Math.round(checksPerMonth);
 };
 
 const getTotalChecksPerMonth = (probeCount: number, frequencySeconds: number) => {
@@ -64,15 +29,8 @@ const getLogsGbPerMonth = (probeCount: number, frequencySeconds: number) => {
   return parseFloat(logsGbPerMonth.toFixed(2));
 };
 
-export const calculateUsage = ({
-  probeCount,
-  checkType,
-  frequencySeconds,
-  useFullMetrics,
-}: ActiveSeriesParams): UsageValues => {
-  const seriesPerCheck = getSeriesPerCheck(checkType, useFullMetrics);
+export const calculateUsage = ({ probeCount, frequencySeconds, seriesPerCheck }: ActiveSeriesParams): UsageValues => {
   const activeSeries = seriesPerCheck * probeCount;
-
   return {
     checksPerMonth: getTotalChecksPerMonth(probeCount, frequencySeconds),
     activeSeries,